:fire: code refactoring
