+++ src/pages/Home/Home.jsx
@@ -1,9 +1,9 @@
-import styled from "styled-components";
-import { Overlap } from "./sections/Main";
-import CurrentHealthGuide from "./sections/Guide";
-import MedicineRecommand from "./sections/Medicine";
-import CommnunityPreview from "./sections/Community/SelectCommunity.jsx";
-import CenterMain from "../../components/CenterMain.jsx";
+import styled from 'styled-components';
+import MainBanner from './sections/MainBanner.jsx';
+import CurrentHealthGuide from './sections/Guide/CurrentDiseaseGuide.jsx';
+import MedicineRecommand from './sections/Medicine/MedicineRecommand.jsx';
+import CommnunityPreview from './sections/Community/SelectCommunity.jsx';
+import CenterMain from '../../components/CenterMain.jsx';
 
 const StyledMain = styled.div`
   background-color: #ffffff;
@@ -18,7 +18,7 @@ const StyledMain = styled.div`
 export default function main() {
   return (
     <StyledMain>
-      <Overlap />
+      <MainBanner />
       <CenterMain>
         <CurrentHealthGuide />
         <MedicineRecommand />
