+++ src/pages/Home/sections/Guide/LifeGuide.jsx
@@ -1,7 +1,7 @@
-import GuideStyle from "./GuideStyle.jsx";
-import { Bold } from "../../../../components/paragraphs.jsx";
+import GuideStyle from './GuideStyle.jsx';
+import { Bold } from '../../../../components/paragraphs.jsx';
 
-export default function OverlapWrapper() {
+export default function LifeGuide() {
   return (
     <GuideStyle>
       <div className="titleWrapper">
