+++ src/pages/MedicineHubPage/index.jsx
@@ -1,10 +1,10 @@
-import { useNavigate } from "react-router-dom";
-import MedicineList from "./MedicineList.jsx";
-import NutrientList from "./NutrientList.jsx";
-import Banner from "../../components/Banner.jsx";
-import SearchBar from "../../components/SearchBar.jsx";
-import CenterMain from "../../components/CenterMain.jsx";
-import RecommandSection from "../../components/RecommandSection.jsx";
+import { useNavigate } from 'react-router-dom';
+import MedicineList from './MedicineList.jsx';
+import NutrientList from './NutrientList.jsx';
+import Banner from '../../components/Banner.jsx';
+import SearchBar from '../../components/SearchBar.jsx';
+import CenterMain from '../../components/CenterMain.jsx';
+import RecommandSection from '../../components/RecommandSection.jsx';
 
 export default function MedicineHubPage() {
   const navigate = useNavigate();
